extract docs to component

+++ app/containers/DevicePage/DevicePage.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react'
 import { useState, useEffect } from 'react'
 import TabSelector from './components/TabSelector'
+import CodeTemplates from './components/CodeTemplates'
 import TextInput from '../../components/TextInput'
 import { Subpage } from '../../model/DevicePage'
 import UconfyLoginApi from '../../services/UconfyLoginApi'
@@ -20,8 +21,10 @@ const DevicePage = (props: Props) => {
   const userData = UconfyLoginApi.getUserData()
 
   const handleDelete = () => {
-    setDeleting(true)
-    props.dispatch(deleteDevice(deviceId))
+    if (confirm('Are you sure you want to delete?')) {
+      setDeleting(true)
+      props.dispatch(deleteDevice(deviceId))
+    }
   }
 
   const handleTabChanged = (currentTab: Subpage) => {
@@ -41,19 +44,34 @@ const DevicePage = (props: Props) => {
 
         {/* Access --------------------------------------------------------------------------------*/}
         {subpage == Subpage.Access && (<>
-          <div className={styles.fieldContainer}>
-            <div className={styles.fieldName}>DeviceID</div>
-            <div className={styles.fieldValue}>
-              <TextInput disabled={true} name={'deviceId'} value={deviceId} onChange={() => {}} />
-            </div>
-          </div>
-          <div className={styles.fieldContainer}>
-            <div className={styles.fieldName}>API Key</div>
-            <div className={styles.fieldValue}>
-              <TextInput disabled={true} name={'apiKey'} value={userData != null ? userData.apiKey : ''} onChange={() => {}} />
-            </div>
-          </div>
-          <button disabled={isDeleting} type="button" className="btn btn-default" onClick={handleDelete}>Delete</button>
+
+          <h3>Access Parameters</h3>
+          <table className={styles.formTable}>
+            <tr>
+              <td>
+                DeviceID
+              </td>
+              <td>
+                 API Key
+              </td>
+            </tr>
+            <tr>
+              <td>
+                <TextInput disabled={true} name={'deviceId'} value={deviceId} onChange={() => {}} />
+              </td>
+              <td>
+                <TextInput disabled={true} name={'apiKey'} value={userData != null ? userData.apiKey : ''} onChange={() => {}} />
+              </td>
+            </tr>
+          </table>
+
+          <CodeTemplates match={props.match} />
+
+          <h3>Delete Device</h3>
+          <p>
+            By deleting the device, your device will no longer be able to push/fetch data.
+          </p>
+          <button disabled={isDeleting} type="button" className="btn btn-danger" onClick={handleDelete}>Delete</button>
         </>)}
 
       </>
